Some low hanging fruit refactors
All changes made are logically equivalent to the existing code, and are
made to make the current code easier to read. The behavior should not be
changed, so also existing bugs should be preserved.
This will probably not be directly used, as the plugin will be more or
less completely rewritten inside of
https://github.com/greenpeace/planet4-master-theme. But I expect parts
can be repurposed.

* Remove dead code.
* Some better naming.
* Some null coalesce-ing.
* Lots of early returns and continue's instead of deep nesting.